I had to make a few changes to the view and feature files. For example:
- The error messages in the view usually didn't match the feature file.
- The feature file included distinct cases for absent and invalid IDs. I thought that was a good idea (which we probably should have done in general, at least in unit tests) so I added an explicit check for missing ID in the view.

Related work items: #147

+++ taskit-backend/tasklists/views.py
@@ -188,14 +188,20 @@ def remove_task_list(request):
         "id": 123
     }
     """
+    if "id" not in request.data:
+        return Response(
+            {"error": "This field is blank."}, status=status.HTTP_400_BAD_REQUEST
+        )
     id = request.data["id"]
     taskLists = TaskList.objects.filter(id=id)
     if taskLists:
         taskList = taskLists.first()
         taskList.delete()
-        return Response({"success": "Task list deleted."}, status=status.HTTP_200_OK)
+        return Response(
+            {"success": "Task list deleted successfully."}, status=status.HTTP_200_OK
+        )
     else:
-        return Response({"error": "Not found"}, status=status.HTTP_404_NOT_FOUND)
+        return Response({"error": "Not found."}, status=status.HTTP_404_NOT_FOUND)
 
 
 # API for editing a task list name
